fixed test so expect_identity.go runs
