+++ helper/resource/testing_new.go
@@ -388,6 +388,7 @@ func runNewTest(ctx context.Context, t testing.T, c TestCase, helper *plugintest
 				fmt.Printf("Step %d/%d Query Output:\n%s\n", stepNumber, len(c.Steps), queryOut)
 				t.Fatalf("Step %d/%d error running query: %s", stepNumber, len(c.Steps), err)
 			}
+
 			err = query.RunQueryChecks(ctx, t, &queryOut, step.ConfigQueryChecks)
 			if err != nil {
 				t.Fatalf("Step %d/%d error running query checks: %s", stepNumber, len(c.Steps), err)
@@ -396,7 +397,6 @@ func runNewTest(ctx context.Context, t testing.T, c TestCase, helper *plugintest
 			fmt.Printf("Step %d/%d Query Output:\n%s\n", stepNumber, len(c.Steps), queryOut)
 
 			continue
-
 		}
 
 		if cfg != nil {
